requirements.txt
asgiref==3.3.0
dj-database-url==0.5.0
Django==3.1.3
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.1
gunicorn==20.0.4
Markdown==3.3.3
numpy==1.19.4
pandas==1.1.4
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
